Move Redirect class to Support\Models namespace
