+++ packages/framework/src/Support/Models/Redirect.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Helpers;
+namespace Hyde\Support\Models;
 
 use Hyde\Hyde;
