+++ packages/framework/tests/Feature/RedirectTest.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Feature;
 
-use Hyde\Helpers\Redirect;
 use Hyde\Hyde;
+use Hyde\Support\Models\Redirect;
 use Hyde\Testing\TestCase;
 
 /**
- * @covers \Hyde\Helpers\Redirect
+ * @covers \Hyde\Support\Models\Redirect
  */
 class RedirectTest extends TestCase
 {
